Attempt to fix review comments.
